requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
numpy==1.23.1
opencv-python
Pillow==9.2.0
